Add a callback to be called once a header field has been read.
Until now, there was no way to detect that a HTTP header field had no value
using this parser.

As a consequence, parsing such header resulted
in having the name of the field concatenated to the name of the next
field (that behavior can be seen in nodejs http-parser).
A test case reproducing the problem as been added.

This patch defines a new header called once the field is completed
(whether or not there has been a value).
It makes it easier to use the parser.

The tests has been updated to use this new callback, the README
has been updated, but links to examples need to be updated.

A patch will be submitted for nodejs to use this new callback.